+++ packages/clients/local-client/src/machine/Machine.ts
@@ -7,7 +7,7 @@
 */
 
 import { CommandStateMachine, CommandStateMachineMode } from "@hive-command/state-machine";
-import { ACTION_TYPES, AssignmentPayload, CommandPayloadItem, CommandProcess, CommandProcessNode, PayloadResponse, ProgramAction } from "@hive-command/data-types";
+import { ACTION_TYPES, AssignmentPayload, CommandPayloadItem, CommandProcess, CommandProcessEdge, CommandProcessNode, PayloadResponse, ProgramAction, ProgramInterlock } from "@hive-command/data-types";
 import { nanoid } from "nanoid";
 import { BusMap } from "./BusMap";
 import { DeviceMap } from "./DeviceMap";
@@ -90,17 +90,17 @@ export class Machine {
 		})
 
 
-		let paths = (flow?.nodes || []).map((action) => {
+		let paths : CommandProcessEdge[] = (flow?.nodes || []).map((action) => {
 
 			const next = flow?.edges?.filter((a) => a.from.id == action.id)
 
 			return (next || []).map((next) => {
 				return {
-					id: nanoid(),
+					id: next.id,
 					source: action?.id, //action.type == "Trigger" ? 'origin' : action.id,
 					target: next?.to?.id,
 					options: {
-						conditions: next?.conditions?.map((cond) => ({
+						conditions: (next?.conditions || []).map((cond) => ({
 							inputDevice: cond.inputDevice.name,
 							inputDeviceKey: cond.inputDeviceKey.key,
 							comparator: cond.comparator,
@@ -114,6 +114,8 @@ export class Machine {
 		}).reduce((prev, curr) => {
 			return prev.concat(curr)
 		}, []).filter((a) => a)
+
+		console.log(JSON.stringify({paths}))
 
 		let subprocs : any = flow?.children?.map((proc) => {
 			return {...this.loadFlow(flow?.children || [], proc.id)}
@@ -194,7 +196,13 @@ export class Machine {
 					interlock: {
 						state: {on: 'true'},
 						// state: {on: true},
-						locks: x.interlocks || []
+						locks: x.interlocks?.map((lock) : ProgramInterlock => ({
+							inputDevice: lock.inputDevice.name,
+							inputDeviceKey: lock.inputDeviceKey.key,
+							comparator: lock.comparator,
+							assertion: lock.assertion,
+							action: lock.action.key
+						})) || []
 					}
 				}
 			}),