+++ packages/core/command-process/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.15-alpha.30](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.29...v1.4.15-alpha.30) (2022-05-22)
+
+**Note:** Version bump only for package @hive-command/process
+
+
+
+
+
+## [1.4.15-alpha.29](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.28...v1.4.15-alpha.29) (2022-05-22)
+
+**Note:** Version bump only for package @hive-command/process
+
+
+
+
+
+## [1.4.15-alpha.28](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.27...v1.4.15-alpha.28) (2022-05-22)
+
+**Note:** Version bump only for package @hive-command/process
+
+
+
+
+
+## [1.4.15-alpha.27](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.26...v1.4.15-alpha.27) (2022-05-22)
+
+**Note:** Version bump only for package @hive-command/process
+
+
+
+
+
 ## [1.4.15-alpha.20](https://github.com/TheTechCompany/HiveCommand/compare/v1.4.15-alpha.19...v1.4.15-alpha.20) (2022-05-21)
 
 **Note:** Version bump only for package @hive-command/process